ajusting merge and modifing code to pass tests

+++ bioptim/optimization/non_linear_program.py
@@ -399,4 +399,4 @@ class NonLinearProgram:
         """
         if node_idx < 0 or node_idx > self.ns:
             return ValueError(f"node_index out of range [0:{self.ns}]")
-        return float(self.tf / self.ns * node_idx)
+        return self.tf / self.ns * node_idx
